Adding file types, cleaning up request counter
I added a few recognized file types, mainly for image files, since I
found a few linked on some sites.

The important change was to the request counter. I turned the code for
creating it into a closure, so that the lines for incrementing it and
updating the HTML could be simplified. I also added an additional
callback to `HTMLDocLoader` called `onComplete`. This callback is
explicitly the last code to be executed during the request, regardless
of whether the request caused an error or went through properly.

This allows me to more cleanly increment and decrement the counter
without covering every return-path in the other two callbacks, or adding
lines to HTMLDocLoader itself (which was quite an egregious violation of
compartmentalization). It will also allow me to, later on, immediately
start preparing the output when the crawling finishes.