Bug Fixes for Banned Strings and Protocol-Relative Link correction, Changing badProtocol to badScheme, and adding presentResults
Some bugs got squashed. The `containsBannedString` function previously
failed to check for banned strings beyond the first one due to incorrect
placement of a return statement. The code for correcting
protocol-relative links (i.e. links whose href attributes start with
`//`) was also fixed so that it no longer mistakenly omitted the colon
";" from "http(s)://".

badProtocol was changed to catch a broader class of erroneous links. As
it turns out, and link whose path contains an unescaped colon ";" before
the first forward slash "/" is interpretted as using a scheme other than
http(s). This includes links using "file://" but also links which have
the incorrect href "internal:node/1234" which I found on a couple of
pages. I don't think there's any legitimate common use-case for these
links, so I log all of them to be checked on a case-by-case basis.

The point when a link is posted to the `allLinks` objects was changed,
so now anchor links are also logged. Null links (those lacking any href
attribute) are still not logged, but I may handle that later.

A couple of other little things were done, such as removing commented
code which has since been replaced by refactored versions.

Finally, I began writing the function `presentResults` which will
present the collected data to the user. So far, all it does is create a
full-screen fixed modal with no meaningful content. I'll keep working on
that.

Ciao!