Recording requests, aborting requests on timeout, Bug fix in scheme recognition, presentResults function and file
Various things gone on here, which is mostly my fault for waiting too
long to make a branch for this.

The main order of business here to properly implement a feature I've had
since basically the beginning of the project, but was broken the entire
time: crawl timeout. The idea was and is to have the script stop if
things take too long (over a minute or so) so that the user isn't
waiting an hour to crawl a massive site. The feature currently seems
buggy (not sure) so when everything is resolved I'll merge this branch
back into master.

I did find a concrete bug in the `classifyLinks()` function, where links
with non-standard schemes wouldn't be recognized if the link didn't
contain a forward-slash `/`. This has been fixed.

(To be clear, the scheme of a url is the part (if any) preceding a colon
`:` which is itself not preceded by a forward-slash. E.g. "http" in
"http://" or "file" in "file:asdf".)

The `presentResults` function has been greatly expanded. I've decided to
just use a style-sheet. This is done by writing all of the CSS text into
a long template-string so that it can be spread across many lines. The
CSS is then placed into a `<style>` tag at the end of the document's
`<head>`. This allows us to take advantage of pseudo-selectors and, in
general, avoid a lot of unnecessary repetition compared to applying CSS
to each element one-by-one. I would even say the modal looks quite good
now! If you just want to see the modal, try pasting the contents of
`presentResults.js` into the console of any site. There's some example
content in their just to show how output currently looks.

Finally, there's a basic demo of using links to download data generated
by the script. This is only available in `classifyLinks.js` at the
moment.