visitLinks rewrite, page unload warning, etc.
I rewrote `visitLinks` as `visitLinks2`. The goal was to remove
`HTMLDocLoader`, and its limitations. Namely, I wanted to be able to
terminate the request as soon as the response-headers were available
(i.e. `httpRequest.readyState === 2`). This allows the script to avoid
wasting time and bandwidth downloading potentially large files which
cannot be parsed for links. This can save a TON of time during a crawl.

Unfortunately, none of the existing callbacks of `HTMLDocLoader` had
access to ready-states other than 4. I decided that, if I had to add a
4th callback-function, it would become too cumbersome and
difficult-to-read to properly serve its purpose of making the request
more concise.

I had already run into this problem partially; I wanted to record all
document requests (`allRequests`), but didn't have access to the
httpRequest object before readyState 4. My original solution was to
return a reference to the httpRequest object from AJAXLoader, which was
then returned by HTMLDocLoader, but this was already cumbersome, and
forced me to record the request only after it had been sent.

Overall, just setting-up and sending the XMLHttpRequest in visitLinks is
only slightly more verbose than HTMLDocLoader, but much more flexible. I
guess I should slightly regret HTMLDocLoader, but it's not the end of
the world. This also makes the flexible AJAXLoader kind of useless for
the same reasons, but leave it for now.

For less-significant changes, I added a line to attach an onbeforeunload
event to the window when the script starts. This partially prevents the
user from accidentally navigating away from the page while the crawler
is running, or while results are displayed. This will help avoid loss of
time and data.

Lastly, I added some code to insert "(INCOMPLETE)" to the download link
if the crawl timed-out. That way, it will be more clear whether any data
is missing, or whether a full crawl actually occurred.