Removed old visitLinks, HTMLDocLoader, Added crawl time
I removed the pre-refactor visitLinks function and renamed `visitLinks2`
back to `visitLinks`. Testing shows that the refactored version does
work properly as far as I can tell.

`HTMLDocLoader` was also removed, as that was factored into the new
`visitLinks`.

I added a timer which starts at the beginning of the script and ends
when the modal is ready to be added to the document. The total time for
the crawl is displayed on the modal next to the "Results:" text. There's
no real practical purpose to this, it's just to quell the user's
concerns about long running times and such. I should note that I added
an extra container and some CSS for this to work properly.

The rest of the changes were mostly just cleanup. I removed some
debugging lines, and some outdated note comments.

Since the goal of this branch has been completed (proper crawl aborting)
I'm going to merge back into master after this commit.