BETA! Removed AJAXLoader, other refactoring, modal search bar
I removed `AJAXLoader` and `textLoader` completely. The remaining AJAX
code has been factored into the new
immediately-executed-function-expression `loadRobotsTxtAndBeginCrawl`.

Other refactoring was also done, though it mostly consisted of moving
some functions around. The start ofthe script is now at the very bottom,
which makes slightly more sense from a reader's perspective. I think.
Maybe not. I don't know.

I added a text-box input to the modal, which allows you to specify which
logging object you want to view/download. I'm planning to change this to
a drop-down in the future, so you don't have to type out the whole name.

Anyways, at this point the script seems to be ready for use by a trained
user besides me, so I'll unceremoniously declare this commit as the
start of the product "beta".