Classify images, related links temporarily excluded, modal style updates
Images are now parsed and classified. To this end, there are several new
logging objects:

* nullImages
* externalImages
* unloadedImages
* bannedStringImages
* absoluteInternalImages

These logging objects are grouped into a separate collection object from
links, so there is now a `linkLoggingObjects` object and a
`imageLoggingObjects` object.

Because images and links are parsed differently when producing output
(since images elements are replaced by their "src" attribute and link
elements are replaced by their "href" attribute), I decided to
concentrate information on those differences into one place: the
`logObjMetaData` array. The only "metadata" currently, though, is what
function to use when converting a logging object to a String (i.e.
JSON). Hopefully, this should make it easy to add new logging
capabilities in the future.

---

There is a small amount of debug code in this commit. It's related to an
issue in Drupal sites, where the "Related Links" block always uses
absolute internal links for some reason.

When that issue is fixed, that piece of code (marked with `//@debug`)
should be removed.

In the meanwhile, though, non-Drupal sites *should* be unaffected,
unless they happen to contain an HTML structure with the same classes.

---

Updates were made to the modal:

* A checkbox was added for changing the format of the outputted JSON.
The previous format (which is still used internally) was made secondary
to a new default format. The new format maps the URL of each page to an
array of *every* link/image href/src on that page which are classified
under the given logging object. This format is much easier for a human
to read and use, since if you find problems you can quickly fix all of
them on a page-by-page basis, rather than having to jump across pages a
bunch.

* A clear button was added which allows you to easily clear the contents
of the object selection textbox

* Border highlights were added to buttons and links so that
tab-navigation is a more viable way to use the modal.

---

Robots.txt parsing was redone. This comes with a fix to a previous bug,
where clauses without spaces following the colon (e.g.
`Disallow:/about`) were not recognized.

---

Other various refactoring occurred, mainly removing some excessive
comments.