Dropdown now works differently on Firefox and Chrome
I added a line to detect whether a webkit browser (such as chrome) or a
gecko browser (such as Firefox) was being used so that could be used to
change the way the suggestion list for the object-selection-textbox
works.

The reason this had to be done is because on Webkit, the label of an
option in that list is displayed alongside the value (both are shown in
the list, with the label being gray-ed out and right-aligned). On Gecko,
only the label is displayed, givent hat it's not missing or empty.

This meant that previously, on Firefox, the dropdown menu would have a
bunch of entries that read only "Empty" rather than showing the name of
the object. So, I just made the label include the object's name along
with an appended `" (Empty)"`, so that the information is still visible
to the user on either browser.