Minify CSS, Refactor checklinks, add highlightElements.js
The CSS was minified, with the original, commented CSS being placed in a
new file: ctrl.js.css

checkLinks was changed in several ways. For one, functions relating to
`BANNED_STRINGS` have been placed in a single object, and a new variable
was added which allows you to force the `isStringBanned` method (new
version of the previous `containedBannedString` function) to return a
true or false value regardless of the parameter.

I made changes in the `visitLinks` function to prevent obscure issues
having to do with references in the logging objects Basically, under
some conditions two objects would get the same array for a given URL,
which would cause changes to one object to be unintentionally reflected
in another. This has been fixed, though it does mean that some data will
now be missed in certain situations. See the new @TODO comment at the
top of the script for more details.

Checking for whether a link is crawlable is now done in visitLinks
rather than in classifyLinks. I'm starting to become more uncertain
about the current line drawn between what these two functions do, but
I'm largely hesitant to move things around much. This change was done
because I feel like deciding not to crawl a link is much more closely
linked to the responsibilities of visitLinks.

This change comes with a change to the handling of robots.txt in
general. I have created a new class constructor function for RobotsTxt.
This collects all of the data and functionality related to robots.txt
handling. This includes requesting the robots.txt file, parsing
patterns, and checking of a page is allowed to be crawled. The object
created to handle this is now passed to visitLinks, making it easy to
spoof or test. The RobotsTxt object also has a flag for ignoring the
file altogether, which causes isUrlAllowed method to simply return true
by default.

The code for starting the crawl has been placed into the `startCrawl`
function, which handles initializing the robotsTxt object. This function
implements new functionality for flags. The function has a string
parameter which can contain flag strings to change certain behavior of
the crawling script without having to modify code directly. There are
currently 3 implemented flags:

1. "-ignoreRobotsTxt" causes the script to skip requesting the
robots.txt file, and simply assume all crawling is allowed. This allows
for convenient crawling of dev sites which typically disallow all
crawler traffic to prevent them from showing up in search engines.

2. "-ignoreBannedStrings" causes the script to behave as though no
strings are banned. This prevents users from having to locate and
manually change the list of banned strings when they want to ignore that
feature. Since links containing banned strings are also not crawled,
this further prevents the script from being stopped on dev sites, if the
devsite name is banned by default.

3. "-ignoreTimeout" and "-ignoreTimer" are equivalent and do exactly the
same thing. They clear the timeout timer which normally halts the script
after a specified amount of time (60 seconds in this commit). If the
timer has been specified needlessly low, or the crawl will likely take
too long, adding this flag prevents the script from being killed early
and lets it run for as long as needed.

All of these flags are case-insensitive, and may have slight variations
to allow for alternate spellings (they are implemented via regular
expressions). The order does not matter. All flags must be preceded by
the hyphen (as written below). You can add spaces and other characters
between flags, but this is not necessary.

To use the flags, simply modify the second parameter of the last line of
the script:

```javascript
startCrawl(robotsTxtHandler, ""); //No flags
startCrawl(robotsTxtHandler, "-ignoreRobotsTxt"); //One flag
startCrawl(robotsTxtHandler,
"-ignoreBannedStrings-ignoreTimer-ignoreRobotsTxt"); //Three flags
```

Lastly, clearChildren function was added in the global scope, though
this is only used in presentResults.