Reworking the rework: A success this time
I became dissatisfied with the increasingly complicated class structure
shown in checkLinks_AltStorage.js (which is currently dysfunctional by
the way; don't try to use it. yetAnotherRework.js is the new good
script). As a result, I decided to start over, and got a nicer, simpler,
and more powerful class structure, as featured in yetAnotherRework.js.

The idea was, primarily, to have an element class which tracks labels
(similar to before) **as well as** manages membership in groups. In
altStorage, you had to handle everything at the level of ElementData
(which grouped together ElementInstances), which got very cumbersome
when trying to get hold of groups based on information about a single
instance. It was also overly complicated to collect instances matching a
given label.

I decided it would be simpler to make the instance-level class the
primary way to manage grouping and labelling, so I restructured
everything into two classes: `ElementRecord` and `RecordGroup`. Both
manage a static field called `LabelTable`, which maps from labels to
arrays of corresponding Records or Groups. Because the classes manage
these tables as you add labels, other code (namely, `presentResults()`)
can easily and reliably take a label and find every `ElementRecord` or
`RecordGroup` with that label.

Additionally, `ElementRecord` manages a static `GroupTable` field, which
allows its constructor to assign it to the appropriate `RecordGroup`
without having to already have a reference to it. This is what allows
the responsibility of group-membership to be delegated to
`ElementRecord`, rather than having to handle things at the group level
as with altStorage.

This will become the new main `checkLinks.js`, as my next commit will be
to rename the file, remove altStorage, and merge with master. I have
accomplished my major goals with this branch, and the script is now
equally functional to the old checkLinks.js, with some additional
features and bug fixes.