New Labels and Bug Fixes
I added 2 new labels: javascriptLink and improperSize. javascriptLinks
are links that use the "javascript:" protocol, which causes the browser
to automatically execute some a script when the user clicks the link.
These are often used in place of buttons and things.

improperSize refers to images which are not being displayed at their
native dimensions, which may be undesirable, especially for very large
images which are resized to be very small. Such large images take up
network resources for no good reason.

I also fixed a bug where the crawler's CSS would sometimes not apply at
all. This was caused by a feature where having multiple stylesheets with
a "name" attribute on a page forces only 1 to work, and all others to be
ignored.

Other changes were mostly miinor refactoring.