Fixed Image Loading and Image Size Checking
I've separated out the process of checking whether an image loads into a
separate ImageLoader class. This class manages requests for new images
and prevents the same image from needing to be requested multiple times.

I've also changed to code around checking whether an image is displayed
at a size besides its native size (the `improperSize` label).
Unfortunately, I've discovered that it cannot be easily determined when
CSS is causing an image to change size, and you cannot directly
determine the displayed size of an image which isn't on the same page.
As a result, the script can only label images as `improperSize` if they
are sized using the `width` and `height` attributes.