Adding Linting, Iframe classification
THis is a mostly stylistic change, But I wanted to ensure a certain code
quality throughout the file. Future edits will abide by this ESLint
file.

I may make changes to the lint file in the future, but I think it's a
good starting-off point, for now.

I also added checking for iframes, mainly to check for ones without
title attributes, which is an accessibility fault. It's also useful to
know, because the crawler cannot check the contents of iframes, so they
must be checked manually for validity.