Bug Fix: Type on markElement
I made a typo in the checkLinks function which caused an erroneous call
to the makeElement function. This would throw an error (due to incorrect
typing) which would halt some requests and prevent proper termination.

I've added a line to the makeElement function to provide a clearer error
message if I make this kind of mistake again.