Adding instruction manual skeleton
There is no contents yet, but I plan to add an equivalent to the
entirety of the PDF file I currently have privately distributed to the
known users of this tool. That way, it will all be in one place, along
with the code, easily updated and viewed, and interactive.