Adding compileTemplate.js, pages->docs
I wrote a templater for the documentation, simply called
"compileTemplate.js". Right now, all it can do are dates (powered by
moment.js) and insert raw file-content. Both of these are used right now
on index.html.template, which compiles to index.html.

The file-insertion feature is particularly useful, as it prevents the
need to duplicate the contents of crawlSite.js into the HTML, which
would otherwise make it very annoying to edit (due to that file being
over 1000 lines long).

I plan to add more features to compileTemplate as necessary, as well as
start using Husky to avoid needing to manually run the compiler before
committing.