Finishing Manual
I've finished adding all needed images and content to the manual, as
well as styling and extra features (namely anchor links) which I'm happy
with.

I've made some tiny adjustments to the compileTemplate.js script as
well, though it was only really adding a template for the HTML comment
at the top of index.html which warns the developer to edit
index.html.template instead and use compileTemplate to create the final
product.