Changing some icons to work with Edge
The heavy check mark and heavy multiplication symbols I was using in the modal before were displaying as big, colored emoji in Edge. I replaced them with lighter symbols that have the same representation. To compensate, I added a text shadow to the multiplication symbol and made the check mark bold.