CSS Changes, Fixing urlRemoveAnchor, and changing link/image/iframe to element labels
I changed some CSS to accomodate for the user agent style sheet of Edge having a few extra rules I didn't anticipate.

I found a bug in edge (incorrectly referenced as "Fixed" here, at the time of this writing: https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/12236493/

This bug causes HTMLAnchorElements ('a' tag DOM nodes) to not have `origin` attributes, though they do in Chrome in Firefox. fortunately, the rest of the location attributes are still present, so I added code to re-construct the origin string from the protocol and host when origin is not available.

I made link, image, and iframe into element labels, when they were previously group labels. Groups are associated with the href/src of the elements they contain. I previously assumed that sites wouldn't have both an image tag and a link tag pointing to the same place, so it made sense before that all elements referring to some URL would be either links or images.

I realize now that this was a bad assumption, partially because of null links/images, which of course refer to the same place (nowhere), as well as iframes, which can most definitely point to the same page as a hyperlink. Previously, when different element types collided in their hrefs/srcs, it was possible for a group to be labelled as both "image" and "link" (or "iframe" for either), which makes little sense.

The logical conclusion was that matching hrefs/srcs does not imply matching element type, so element type should be associated with individual elements, not groups.