Fixing Documentation, Adding file-escape-html to compileTemplate
I realized a `<li>` in a *comment* in the crawlSite.js script was messing up the rendered HTML when the file text was embedded into `index.html.template`, so I added a new module `escape-html` and a complementary new directive `file-escape-html:` to `compileTemplate.js`.

This is a sort of stop-gap solution, though. Right now the code for the `file-escape-html` directive is almost entirely duplicated from the `file` directive. In the future, I should implement some sort of pipeline construct to allow slight modification of embedded text to mitigate this code duplication.

I'll tentatively propose the syntax `file:my_file_name.ext|escape-html`, though I have no current roadmap for implementing this.