Coyping changes from `crlr-suggestionList.js.css` to `crlr.js.css`, ading CSS-Select Flag
I had forgotten to delete `crlr-suggestionList.js.css` and copy over the changes to `crlr.js.css` before, so I fixed that now. I've also made various changes to the way labels are displayed by allowing each label to have custom display settings. This change was part of the implementation of the new `-css-select` flag, which allows the user to specify a CSS selector which runs against every page crawled. Selected elements are collected into the `user-selected` label.

I still have to update the documentation to reflect this