Allow inverted-output changing when the label field is clear or invalid, removed debugging statements
Previously, if the value of the label textbox was not a valid label (e.g. it was cleared or the user was in the middle of typing), changing the inverted-output-format checkbox would not actually affect the output. This has been fixed by having the input handler keep track of the currently displayed label.