Changing Minimize Icon, Suggestion List optimizations and improvements
I changed the minimize icon to use a rectangular div instead of the actual Unicode character, as it wasn't displaying correctly on some windows 7 PCs.

I made some further performance optimizations tot he suggestion list. It ill no longer update the output if it is currently displaying all suggestions and would continue to display all suggestions after the update. Previously, the list would be cleared and re-filled with each character typed.

I also added a new button next to the input element which is symbolized by a downward-pointing black triangle. Clicking this button will open the usggestion list, even if the input element was not already focused. This replicates the functionality of the default input/datalist elements in HTML.

Finally, I replaced a bunch of double-quotes with single-quotes to better fit my own, personal, very special, and superior semantic distinction between them. Single quotes should be used anywhere that a string is treated like a key, symbol, or internal value. Double quotes should be used anywhere a string is part of some content to be displayed to the user.

Note that there's a few places I wasn't sure which was more appropriate, like for setting labels and such, but it's kind of a meager distinction anyways and I don't care that much to solidify things.