Lexing Progress, Filter profiling, userSelected grouping, crawlSite_SLOW
I made some progress in fleshing out the mechanics of lexing and executing label queries. I've done some fairly extensive profiling in order to determine the best approach to use.

As of now, I think the best general solution is simply to run a filter over every ElementRecord (which can now be easily gotten via ElementRecord.ALL. In general, I've found this to be either comparable to or faster than an approach based on Set unions, intersections, and negations. I suspect this is due to the lack of intermediate array/Set construction and collection.

The main exception to this lies in fairly specific special cases, such as queries which consist of intersections ("and" operations, e.g. "link and internal and absolute"), since those can be more quickly calculated using the `LabelTable`s before filtering. Depending on whether performance issues show up when this feature is complete, I may add special-case optiimizations to cover such scenarios.

I changed the way groupNames are determined for user-selected elements. Now, elements with `href` or `src` attributes are added to their corresponding navigation groups, and an element's `tagName` is used as the fallback. This is allows for user-selected links/images/iframes to get group labels. They do not get element labels, however. I would like to fix this, but not this commit.

I added a very unstable throttled/buffered version of crawlSite, which uses the `buffer` function from my Utilities repository. This file was used for a site which was causing Gateway errors due to the high rate of requests, and this script was my attempt at slowing things down discretely. It caused some bugs, though, so I won't recommend using it.