Adding class RecordCache
I've added a new class `RecordCache` to resolve the issue I presented in the previous commit, which was that if the user-selected flag was supplied with a selector that matched links/images/iframes, user-selected elements would previously be given separate ElementRecords from the ones generated from other `classify` functions.

Now, when each page is crawled, a `RecordCache` instance is created for that page/URL. This instance has a `getFor` method to which an element is passed. If that element already in the cache, the corresponding ElementRecord is returned. If not, a new ElementRecord is created for that element. The result is that, per page, there is now a 1-to-1 mapping from HTML elements to ElementRecords across all `classify` functions , which means labels are always shared.