evaluateBoolExpression
Added a function (but not any functionality) which models how I think the evaluator should actually work. The current evaluateSetExpression appears to be very slow compared to iterating over similarly-sized data and using a filter to get the matching elements. I suspect this is due to the many intermediate arrays/Sets that are constructed by evaluateSetExpression. For a filter method, however, we need to be able to evaluate the query on a per-ElementRecord basis, which necessitates this new function.