Slight change to http-httpsError, evalBool refactor
http-httpsError now contains external links with mismatched protocols. Previously, it only contained internal links.

I've refactored the `evaluateBoolExpression` function from `lexTest.js` in the new file `evalBoolRefactor.js`. The new function `makeRecordFilter` is much simpler and shorter (roughly 40 lines shorter), and is probably what I'l use in the final version of this feature.

Much of this simplification is thanks to the new feature I've made to `linearLex`: parenthetical operator support. The grammar can now specify an `open` and `close` property on token types, and will keep a stack of opening parentheticals as it reads through the given expression string. If the parentheticals lack a match, or are nested in incorrect order, an error will be thrown.

This new feature makes linearLex much more powerful, and allows it to succinctly handle even fairly complex grammars such as XML.