Improving interopperability between complex queries and the suggestion list (TokenizedInputElement)
I've made it so that the suggestion list will only make suggestions for and replacements to the currently selected token(s) in the query.

This was accomplished primarily through the introduction of the TokenizedInputElement class, which generates and stores a tokenization of the wrapped input element's value. It provides utility functions for identifying and replacing the selected tokens within that value. These functions are hooked into the already-implemented optional parameters to `makeAutoCompleteList` which override the function's behavior for reading and writing the input element's value.

Moreover, I have introduced another class, `InputSelectionTracker`, which essentially allows for the script to listen to changes to the selection of an input element as though it were an event. Indeed, this is proposed functionality for the `InputElement selectionchange` event, but it has not been implemented any stable browser release. Instead, the class listens to the widely implemented `document selectionchange` event and tracks the `selectionStart/End` properties of each wrapped input element in order to alert the corresponding listeners to changes.

Other changes include altering the outline color of the textbox when the query is invalid in terms of format, and various bug fixes made to linearLex.

Next commit will probably be documentation, followed by merging w/ master.